fix: Move prisma into services, adjust types to support env variables and global services, update apollo server to v3, prepare for material-ui v5 migration, move generated types to new directoriy
