+++ codegen.yml
@@ -3,7 +3,7 @@ overwrite: true
 schema: 'http://localhost:3000/api/graphql'
 documents: '{components,pages}/**/*.{ts,tsx}'
 generates:
-    gen/graphql-types.tsx:
+    types/gen/graphql-types.tsx:
         config:
             dedupeOperationSuffix: true
             withHooks: true
