+++ components/dialog/ToolDialog.tsx
@@ -3,7 +3,7 @@ import { Tool } from '@prisma/client';
 import gql from 'graphql-tag';
 import React, { ReactElement } from 'react';
 import { Controller, useForm } from 'react-hook-form';
-import { useCreateToolMutation } from '../../gen/graphql-types';
+import { useCreateToolMutation } from '../../types/gen/graphql-types';
 import { QUERY_TOOLS } from '../../pages';
 
 interface Props {
