--- gen/fragmentTypes.json
@@ -1,3 +1,0 @@
-{
-  "possibleTypes": {}
-}
