+++ package.json
@@ -47,11 +47,11 @@
         "@babel/plugin-transform-typescript": "^7.15.0",
         "@babel/preset-env": "^7.15.0",
         "@babel/preset-typescript": "^7.15.0",
-        "@graphql-codegen/cli": "^2.0.1",
-        "@graphql-codegen/fragment-matcher": "^3.0.0",
-        "@graphql-codegen/typescript": "^2.0.0",
-        "@graphql-codegen/typescript-operations": "^2.0.1",
-        "@graphql-codegen/typescript-react-apollo": "^3.0.0",
+        "@graphql-codegen/cli": "^2.1.0",
+        "@graphql-codegen/fragment-matcher": "^3.1.0",
+        "@graphql-codegen/typescript": "^2.1.0",
+        "@graphql-codegen/typescript-operations": "^2.1.0",
+        "@graphql-codegen/typescript-react-apollo": "^3.1.0",
         "@testing-library/dom": "^8.1.0",
         "@testing-library/jest-dom": "^5.14.1",
         "@testing-library/react": "^12.0.0",
