+++ types/global.d.ts
@@ -1,0 +1,17 @@
+import { PrismaClient } from '../services/prisma';
+
+declare global {
+    namespace NodeJS {
+        interface ProcessEnv {
+            NODE_ENV: 'development' | 'production';
+            PORT?: string;
+            DATABASE_URL: string;
+        }
+
+        interface Global {
+            prisma: PrismaClient;
+        }
+    }
+}
+
+export {};
